+++ packages/fuselage/src/components/ProgressBar/spec.tsx
@@ -5,6 +5,6 @@ import { ProgressBar } from '../..';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
-  ReactDOM.render(<ProgressBar />, div);
+  ReactDOM.render(<ProgressBar percentage={0} />, div);
   ReactDOM.unmountComponentAtNode(div);
 });
