+++ packages/fuselage/src/components/ProgressBar/styles.scss
@@ -2,18 +2,6 @@
 @use '../../styles/lengths.scss';
 @use '../../styles/typography.scss';
 
-$progress-bar-color-primary: theme(
-  'progress-bar-color-primary',
-  colors.primary(200)
-);
-$progress-bar-color-success: theme(
-  'progress-bar-color-success',
-  colors.success(200)
-);
-$progress-bar-color-error: theme(
-  'progress-bar-color-error',
-  colors.danger(200)
-);
 $progress-bar-color-shine: theme('progress-bar-color-shine', colors.surface());
 $progress-bar-color-background: theme(
   'progress-bar-color-background',
@@ -38,33 +26,21 @@ $progress-bar-color-background: theme(
 
   border-radius: lengths.border-radius(2);
 
-  &--type-primary {
-    background-color: $progress-bar-color-primary;
+  &-complete::before {
+    position: absolute;
+    top: 0;
+    bottom: 0;
+    left: 0;
+    right: 0;
 
-    &::before {
-      position: absolute;
-      top: 0;
-      bottom: 0;
-      left: 0;
-      right: 0;
+    width: inherit;
 
-      width: inherit;
+    content: '';
+    animation: rcx-progress-bar__animation 2s ease-out infinite;
 
-      content: '';
-      animation: rcx-progress-bar__animation 2s ease-out infinite;
-
-      opacity: 0;
-      border-radius: lengths.border-radius(2);
-      background: $progress-bar-color-shine;
-    }
-  }
-
-  &--type-success {
-    background-color: $progress-bar-color-success;
-  }
-
-  &--type-error {
-    background-color: $progress-bar-color-error;
+    opacity: 0;
+    border-radius: lengths.border-radius(2);
+    background: $progress-bar-color-shine;
   }
 }
